Move ChartSettings to Dashboard and Question variants (#49220)
