+++ frontend/src/metabase-types/store/qb.ts
@@ -1,4 +1,5 @@
 import type { QueryModalType } from "metabase/query_builder/constants";
+import type { Widget } from "metabase/visualizations/components/ChartSettings/types";
 import type {
   Card,
   DashboardId,
@@ -11,6 +12,10 @@ import type {
 export type QueryBuilderMode = "view" | "notebook" | "dataset";
 export type DatasetEditorTab = "query" | "metadata";
 export type QueryBuilderQueryStatus = "idle" | "running" | "complete";
+export type InitialChartSettingState = {
+  section?: string | null;
+  widget?: Widget | null;
+};
 
 export type ForeignKeyReference = {
   status: number;
@@ -32,7 +37,7 @@ export interface QueryBuilderUIControls {
   isShowingSnippetSidebar: boolean;
   isShowingTimelineSidebar: boolean;
   isNativeEditorOpen: boolean;
-  initialChartSetting: null;
+  initialChartSetting: InitialChartSettingState;
   isShowingRawTable: boolean;
   queryBuilderMode: QueryBuilderMode | false;
   previousQueryBuilderMode: boolean;
