+++ frontend/src/metabase/dashboard/components/DashCard/DashCardActionsPanel/DashCardActionButton/index.ts
@@ -1,0 +1,1 @@
+export * from "./DashCardActionButton";
