+++ frontend/src/metabase/visualizations/components/ChartSettings/ChartSettingsFooter/index.ts
@@ -1,0 +1,1 @@
+export * from "./ChartSettingsFooter";
