+++ frontend/src/metabase/visualizations/components/ChartSettings/ChartSettingsVisualization/index.ts
@@ -1,0 +1,1 @@
+export * from "./ChartSettingsVisualization";
