+++ frontend/src/metabase/visualizations/components/settings/ChartNestedSettingsSeriesMultiple.unit.spec.js
@@ -2,7 +2,7 @@
 import userEvent from "@testing-library/user-event";
 
 import { renderWithProviders, screen, within } from "__support__/ui";
-import { ChartSettings } from "metabase/visualizations/components/ChartSettings";
+import { DashboardChartSettings } from "metabase/visualizations/components/ChartSettings";
 import registerVisualizations from "metabase/visualizations/register";
 
 registerVisualizations();
@@ -35,7 +35,7 @@ const setup = (seriesDisplay, numberOfSeries = 1, changeSeriesName = false) => {
     .fill(1)
     .map((s, index) => getSeries(seriesDisplay, index, changeSeriesName));
   return renderWithProviders(
-    <ChartSettings
+    <DashboardChartSettings
       series={series}
       initial={{ section: "Display" }}
       isDashboard={true}
