+++ frontend/src/metabase/visualizations/visualizations/BarChart.unit.spec.tsx
@@ -1,6 +1,6 @@
 import { createMockMetadata } from "__support__/metadata";
 import { renderWithProviders, screen } from "__support__/ui";
-import { ChartSettings } from "metabase/visualizations/components/ChartSettings";
+import { QuestionChartSettings } from "metabase/visualizations/components/ChartSettings";
 import registerVisualizations from "metabase/visualizations/register";
 import Question from "metabase-lib/v1/Question";
 import type { Series } from "metabase-types/api";
@@ -24,11 +24,10 @@ interface SetupProps {
 
 const setup = ({ series, question }: SetupProps) => {
   renderWithProviders(
-    <ChartSettings
+    <QuestionChartSettings
       series={series}
       question={question}
       initial={{ section: "Data" }}
-      noPreview
     />,
   );
 };
