+++ frontend/src/metabase/visualizations/visualizations/PivotTable/PivotTable.unit.spec.js
@@ -4,7 +4,7 @@ import { useState } from "react";
 
 import { createMockMetadata } from "__support__/metadata";
 import { render, screen } from "__support__/ui";
-import { ChartSettings } from "metabase/visualizations/components/ChartSettings";
+import { QuestionChartSettings } from "metabase/visualizations/components/ChartSettings";
 import registerVisualizations from "metabase/visualizations/register";
 import Question from "metabase-lib/v1/Question";
 import { createMockColumn } from "metabase-types/api/mocks";
@@ -61,7 +61,7 @@ const setup = () => {
     };
 
     return (
-      <ChartSettings
+      <QuestionChartSettings
         onChange={onChange}
         series={[
           {
@@ -93,7 +93,6 @@ const setup = () => {
           },
         ]}
         initial={{ section: "Data" }}
-        noPreview
         question={question}
       />
     );
